fix issue with response status in the API requests
