+++ src/components/AddBook.js
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { v4 as uuidv4 } from 'uuid';
-import { addBookToApi } from '../redux/books/books';
+import { addBookToApi } from '../redux/books/middlewares';
 
 const AddBook = () => {
   const initialState = { title: '', author: '' };
